Both the player and AI are now both bizarre ninjas.
The "dash" towards the ball is caused by the strike zone area being too large. We re-calculate towards the middle of it as soon as our player could possibly reach the edge.

What we need to do is determine the exact pose that a character will hit before it hits it, based on look-ahead hit in a smaller strike zone. Once we're anywhere in there, we smoothly interpolate to the precise pose value. The strike zone's center must be that ideal position!